Extract naming convention verification
